storage/fs: move encryption-at-rest from engineccl package
Move the code supporting encryption-at-rest out of the ccl/ tree and into
storage/fs.

Epic: none
Release note: none
