+++ pkg/BUILD.bazel
@@ -102,7 +102,6 @@ ALL_TESTS = [
     "//pkg/ccl/sqlproxyccl/tenant:tenant_test",
     "//pkg/ccl/sqlproxyccl/throttler:throttler_test",
     "//pkg/ccl/sqlproxyccl:sqlproxyccl_test",
-    "//pkg/ccl/storageccl/engineccl:engineccl_test",
     "//pkg/ccl/storageccl:storageccl_test",
     "//pkg/ccl/telemetryccl:telemetryccl_test",
     "//pkg/ccl/testccl/authccl:authccl_test",
@@ -1093,8 +1092,6 @@ GO_TARGETS = [
     "//pkg/ccl/sqlproxyccl/throttler:throttler_test",
     "//pkg/ccl/sqlproxyccl:sqlproxyccl",
     "//pkg/ccl/sqlproxyccl:sqlproxyccl_test",
-    "//pkg/ccl/storageccl/engineccl:engineccl",
-    "//pkg/ccl/storageccl/engineccl:engineccl_test",
     "//pkg/ccl/storageccl:storageccl",
     "//pkg/ccl/storageccl:storageccl_test",
     "//pkg/ccl/telemetryccl:telemetryccl_test",
