--- pkg/ccl/storageccl/engineccl/.gitignore
@@ -1,6 +1,0 @@
-# Do not add environment-specific entries here (see the top-level .gitignore
-# for reasoning and alternatives).
-
-# Old benchmark data.
-mvcc_data
-mvcc_data_*
