+++ pkg/storage/fs/BUILD.bazel
@@ -4,9 +4,11 @@ go_library(
     name = "fs",
     srcs = [
         "category.go",
+        "ctr_stream.go",
         "encryption_at_rest.go",
         "file_registry.go",
         "fs.go",
+        "key_manager.go",
         "min_version.go",
         "sticky_vfs.go",
         "temp_dir.go",
@@ -38,13 +40,20 @@ go_library(
         "@com_github_cockroachdb_pebble//vfs/atomicfs",
         "@com_github_cockroachdb_pebble//vfs/vfstest",
         "@com_github_cockroachdb_redact//:redact",
+        "@com_github_gogo_protobuf//proto",
+        "@com_github_lestrrat_go_jwx_v2//jwa",
+        "@com_github_lestrrat_go_jwx_v2//jwk",
     ],
 )
 
 go_test(
     name = "fs_test",
     srcs = [
+        "ctr_stream_test.go",
+        "encryption_at_rest_test.go",
         "file_registry_test.go",
+        "key_manager_test.go",
+        "main_test.go",
         "sticky_vfs_test.go",
         "temp_dir_test.go",
     ],
@@ -61,9 +70,15 @@ go_test(
         "//pkg/util/debugutil",
         "//pkg/util/leaktest",
         "//pkg/util/log",
+        "//pkg/util/protoutil",
+        "//pkg/util/timeutil",
         "@com_github_cockroachdb_datadriven//:datadriven",
+        "@com_github_cockroachdb_errors//:errors",
         "@com_github_cockroachdb_errors//oserror",
         "@com_github_cockroachdb_pebble//vfs",
+        "@com_github_cockroachdb_pebble//vfs/atomicfs",
+        "@com_github_cockroachdb_pebble//vfs/errorfs",
+        "@com_github_gogo_protobuf//proto",
         "@com_github_kr_pretty//:pretty",
         "@com_github_stretchr_testify//require",
     ],
