+++ pkg/storage/fs/ctr_stream.go
@@ -3,7 +3,7 @@
 // Use of this software is governed by the CockroachDB Software License
 // included in the /LICENSE file.
 
-package engineccl
+package fs
 
 import (
 	"context"
