+++ pkg/storage/fs/ctr_stream_test.go
@@ -3,7 +3,7 @@
 // Use of this software is governed by the CockroachDB Software License
 // included in the /LICENSE file.
 
-package engineccl
+package fs
 
 import (
 	"bytes"
@@ -89,10 +89,10 @@ func encryptManySubBlocks(
 }
 
 // Running non-fips mode:
-// ./dev test pkg/ccl/storageccl/engineccl -f CTRStreamDataDriven  --rewrite --stream-output
+// ./dev test pkg/storage/fs -f CTRStreamDataDriven  --rewrite --stream-output
 // Running fips mode:
-// ./dev test-binaries --cross=crosslinuxfips pkg/ccl/storageccl/engineccl && mkdir -p fipsbin && tar xf bin/test_binaries.tar.gz -C fipsbin && docker run -v
-// $PWD/fipsbin:/fipsbin -it redhat/ubi9 bash -c 'cd /fipsbin/pkg/ccl/storageccl/engineccl/bin && ./run.sh -test.run CTRStreamDataDriven'
+// ./dev test-binaries --cross=crosslinuxfips pkg/storage/fs && mkdir -p fipsbin && tar xf bin/test_binaries.tar.gz -C fipsbin && docker run -v
+// $PWD/fipsbin:/fipsbin -it redhat/ubi9 bash -c 'cd /fipsbin/pkg/storage/fs/bin && ./run.sh -test.run CTRStreamDataDriven'
 func TestCTRStreamDataDriven(t *testing.T) {
 	defer leaktest.AfterTest(t)()
 	for _, impl := range []string{"v1", "v2"} {
@@ -338,10 +338,10 @@ func TestFileCipherStreamCreator(t *testing.T) {
 }
 
 // Running non-fips mode:
-// ./dev bench pkg/ccl/storageccl/engineccl -f FileCipherStream --stream-output --ignore-cache
+// ./dev bench pkg/storage/fs -f FileCipherStream --stream-output --ignore-cache
 // Running fips mode (be sure to look for fips=true in the output):
-// ./dev test-binaries --cross=crosslinuxfips pkg/ccl/storageccl/engineccl && mkdir -p fipsbin && tar xf bin/test_binaries.tar.gz -C fipsbin && docker run -v
-// $PWD/fipsbin:/fipsbin -it redhat/ubi9 /fipsbin/pkg/ccl/storageccl/engineccl/bin/engineccl_test -test.run '^$' -test.bench FileCipherStream
+// ./dev test-binaries --cross=crosslinuxfips pkg/storage/fs && mkdir -p fipsbin && tar xf bin/test_binaries.tar.gz -C fipsbin && docker run -v
+// $PWD/fipsbin:/fipsbin -it redhat/ubi9 /fipsbin/pkg/storage/fs/bin/fs_test -test.run '^$' -test.bench FileCipherStream
 func BenchmarkFileCipherStream(b *testing.B) {
 	isFips := fips140.Enabled()
 	for _, impl := range []string{"v1", "v2"} {
