+++ pkg/storage/fs/file_registry.go
@@ -29,19 +29,28 @@ import (
 	"github.com/cockroachdb/redact"
 )
 
-// DefaultNumOldFileRegistryFiles is the default number of old registry files
+// defaultNumOldFileRegistryFiles is the default number of old registry files
 // kept for debugging. Production callers should use this to initialize
 // FileRegistry.NumOldRegistryFiles. The value of two is meant to slightly align
 // the history with what we keep for the Pebble MANIFEST. We keep one Pebble
 // MANIFEST, which is also 128MB in size, however the entry size per file in the
 // MANIFEST is smaller (based on some unit test data it is about half), so we
 // keep double the history for the file registry.
-var DefaultNumOldFileRegistryFiles = envutil.EnvOrDefaultInt(
+var defaultNumOldFileRegistryFiles = envutil.EnvOrDefaultInt(
 	"COCKROACH_STORE_NUM_OLD_FILE_REGISTRY_FILES", 2)
 
-// CanRegistryElideFunc is a function that returns true for entries that can be
-// elided instead of being written to the registry.
-var CanRegistryElideFunc func(entry *enginepb.FileEntry) bool
+// elidePlaintext is a function that returns true for entries that can be elided
+// instead of being written to the registry.
+func elidePlaintext(entry *enginepb.FileEntry) bool {
+	if entry == nil {
+		return true
+	}
+	settings := &enginepb.EncryptionSettings{}
+	if err := protoutil.Unmarshal(entry.EncryptionSettings, settings); err != nil {
+		return false
+	}
+	return settings.EncryptionType == enginepb.EncryptionType_Plaintext
+}
 
 const defaultSoftMaxRegistrySize = 128 << 20 // 128 MB
 
@@ -75,6 +84,10 @@ type FileRegistry struct {
 	// exceeded the registry may consider rolling over to a new file. Defaults
 	// to defaultSoftMaxRegistrySize.
 	SoftMaxSize int64
+	// CanElideEntry is a function that returns true for entries that can be
+	// elided (e.g., they're plaintext and don't need to be stored in the
+	// registry). It's configurable for ease of testing.
+	CanElideEntry func(entry *enginepb.FileEntry) bool
 
 	// Implementation.
 
@@ -314,9 +327,9 @@ func (r *FileRegistry) maybeElideEntries(ctx context.Context) error {
 			panic(errors.AssertionFailedf("entry disappeared from map"))
 		}
 
-		// Some entries may be elided. This is used within
-		// ccl/storageccl/engineccl to elide plaintext file entries.
-		if CanRegistryElideFunc != nil && CanRegistryElideFunc(entry) {
+		// Some entries may be elided. This is used by encryption-at-rest to
+		// elide plaintext file entries.
+		if r.CanElideEntry != nil && r.CanElideEntry(entry) {
 			batch.DeleteEntry(filename)
 			continue
 		}
