+++ pkg/storage/fs/main_test.go
@@ -3,6 +3,6 @@
 // Use of this software is governed by the CockroachDB Software License
 // included in the /LICENSE file.
 
-package engineccl
+package fs
 
 //go:generate ../../../util/leaktest/add-leaktest.sh *_test.go
