+++ pkg/storage/fs/testdata/ctr_stream
